requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
attrs==25.3.0
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist==1.7.0
greenlet==3.2.4
gunicorn==21.2.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
propcache==0.3.2
PyJWT==2.10.1
pyotp==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.0
qrcode==7.4.2
requests==2.32.4
SQLAlchemy==2.0.42
twilio==9.0.0
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
yarl==1.20.1
cryptography
Flask-Talisman
Flask-Limiter
passlib